Add recursive safety timeout | Lower timeout
Since we now have a recursive timeout check we can lower the wait time and just have it call the the function again if the canvas hasn't been sized yet. Still a little flicker that should be dealt with if it has not wait multiple cycles to resize. Don't do this a future game restarts. Only first load.